test(nifs): prepare for nifs tests in separate file
